[Refactor] Change project structure
