+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.CampingMate"
         tools:targetApi="31">
         <activity
-            android:name=".ui.WebViewActivity"
+            android:name=".WebViewActivity"
             android:exported="false" />
         <activity
             android:name=".NewActivity"
