+++ app/src/main/java/com/brandon/campingmate/PostListViewType.kt
@@ -1,0 +1,14 @@
+package com.brandon.campingmate
+
+enum class PostListViewType {
+    POSTITEM,
+    LOADING,
+    UNKNOWN,
+    ;
+
+    companion object {
+        fun from(ordinal: Int): PostListViewType = entries.find {
+            it.ordinal == ordinal
+        } ?: UNKNOWN
+    }
+}
