+++ app/src/main/java/com/brandon/campingmate/SearchListAdapter.kt
@@ -6,7 +6,6 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.brandon.campingmate.data.CampModel
 import com.brandon.campingmate.databinding.ItemBigCampBinding
 import com.bumptech.glide.Glide
