+++ app/src/main/java/com/brandon/campingmate/ViewPager2Adapter.kt
@@ -4,11 +4,6 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.brandon.campingmate.ui.BoardFragment
-import com.brandon.campingmate.ui.ChatFragment
-import com.brandon.campingmate.ui.HomeFragment
-import com.brandon.campingmate.ui.MapFragment
-import com.brandon.campingmate.ui.ProfileFragment
 
 class ViewPager2Adapter(
     fragmentManager: FragmentManager,
