--- app/src/main/java/com/brandon/campingmate/data/PostListViewType.kt
@@ -1,14 +1,0 @@
-package com.brandon.campingmate.data
-
-enum class PostListViewType {
-    POSTITEM,
-    LOADING,
-    UNKNOWN,
-    ;
-
-    companion object {
-        fun from(ordinal: Int): PostListViewType = entries.find {
-            it.ordinal == ordinal
-        } ?: UNKNOWN
-    }
-}
