--- app/src/main/java/com/brandon/campingmate/data/PostModel.kt
@@ -1,16 +1,0 @@
-package com.brandon.campingmate.data
-
-import com.google.firebase.Timestamp
-
-data class PostModel(
-    val author: String? = null,
-    val authorId: String? = null,
-    val authorProfileImageUrl: String? = null,
-    val title: String? = null,
-    val content: String? = null,
-    val imageUrlList: List<String>? = null,
-    val timestamp: Timestamp? = null //
-)
-
-//    val timestamp: Timestamp? = Timestamp.now() // Firestore의 Timestamp 사용
-
