+++ app/src/main/res/layout/activity_web_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.WebViewActivity">
+    tools:context=".WebViewActivity">
 
     <WebView
         android:id="@+id/wv_webview"
