+++ app/src/main/res/navigation/bottom_navigation.xml
@@ -8,27 +8,27 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.brandon.campingmate.ui.HomeFragment"
+        android:name="com.brandon.campingmate.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/boardFragment"
-        android:name="com.brandon.campingmate.ui.BoardFragment"
+        android:name="com.brandon.campingmate.BoardFragment"
         android:label="fragment_board"
         tools:layout="@layout/fragment_board" />
     <fragment
         android:id="@+id/chatFragment"
-        android:name="com.brandon.campingmate.ui.SearchFragment"
+        android:name="com.brandon.campingmate.SearchFragment"
         android:label="fragment_chat"
         tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/mapFragment"
-        android:name="com.brandon.campingmate.ui.MapFragment"
+        android:name="com.brandon.campingmate.MapFragment"
         android:label="fragment_map"
         tools:layout="@layout/fragment_map" />
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.brandon.campingmate.ui.ProfileFragment"
+        android:name="com.brandon.campingmate.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
